requirements.txt
cohere
google-generativeai
pillow
pymupdf
chromadb
langchain
langchain-google-genai
gradio